+++ sdks/ruby/lib/satusehat_sdk/models.rb
@@ -1,0 +1,5 @@
+module Satusehat
+    module IntegrationSdk
+        FhirResponse = Struct.new(:status_code, :body, :headers, keyword_init: true)
+    end
+end